Fix CLI lint errors
