+++ typescript/cli/src/tests/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "import/no-cycle": ["off"]
+  }
+}
